Revised Version of Traditional Chinese Translation
Reason for changing filename:
the Chinese character 繁 not only means "complex/complicated", but also carries a negative connotation of something similar to "perplexing" in English. This should be changed for the sake of impartiality.
Modification:
I changed the wording of many phrases to make it sound more natural for a native Chinese speaker. (Chinese is hard!) However, I have very little understanding about logical/philosophical aspects of lojban, so I am not able to translate the part regarding that for now.